refactor(scripts): introduce learning_kit package and move learning modules (BREAKING CHANGE)
This commit introduces a major refactor by **adding a new package**
`learning_kit/` and **moving key modules into it**. This change has
restructured the learning logic and affects the entire `script/` directory.

### Changes:
- Added a new package `learning_kit/` to organize all utilities related to learning logic.
- Moved the following modules into `learning_kit/`:
  - `inference.py`
  - `neuralnet.py`
  - `metrics.py`
- Updated import dependencies across the `scripts/` directory to follow
-  the new package structure.

### BREAKING CHANGE:
This refactor induces breaking changes as the file structure has been modified,
requiring updates to import statements throughout the codebase.

### Why this is important:
This restructuring improves organization and maintainability
by grouping learning-related modules under a unified package.
It also sets a clearer structure for future development within the learning logic.
